requirements.txt
Django>=5.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
graphene-django>=3.1.5
django-cors-headers>=4.3